fix: 퀴즈 보기에 option_id 추가 -> 전체 기능 로직 수정

+++ backend/quiz/src/main/java/com/boindang/quiz/domain/QuizSolvedHistory.java
@@ -24,11 +24,11 @@ public class QuizSolvedHistory {
 
 	private boolean isCorrect;
 
-	private Long selectedOptionId;
+	private int selectedOptionId;
 
 	private LocalDateTime solvedAt;
 
-	public QuizSolvedHistory(Long userId, Quiz quiz, boolean isCorrect, Long selectedOptionId) {
+	public QuizSolvedHistory(Long userId, Quiz quiz, boolean isCorrect, int selectedOptionId) {
 		this.userId = userId;
 		this.quiz = quiz;
 		this.isCorrect = isCorrect;
