+++ backend/quiz/src/main/java/com/boindang/quiz/presentation/dto/request/AnswerRequest.java
@@ -2,5 +2,5 @@ package com.boindang.quiz.presentation.dto.request;
 
 public record AnswerRequest(
 	Long quizId,
-	Long selectedOptionId
+	int selectedOptionId
 ) {}
